requirements.txt
fastapi
uvicorn
pyngrok
langchain
langchain-community 
twilio
python-decouple
python-multipart
google-cloud-dialogflow
dialogflow
